+++ src/app/auth/user-auth.component.html
@@ -1,0 +1,60 @@
+<div class="container my-5">
+  <!-- SIGN UP -->
+  <div *ngIf="!showLogin">
+    <div class="card">
+      <div class="card-header">
+        <h1>User Signup</h1>
+      </div>
+      <div class="card-body">
+        <form class="common-form" [formGroup]="signupForm" (ngSubmit)="signUp()">
+
+          <input type="text" class="form-input" formControlName="name" placeholder="Enter User Name" />
+          <p *ngIf="signupForm.get('name')?.touched && signupForm.get('name')?.invalid" class="invalid-input">
+            Please enter a valid name
+          </p>
+
+          <input type="email" class="form-input" formControlName="email" placeholder="Enter User Email" />
+          <p *ngIf="signupForm.get('email')?.touched && signupForm.get('email')?.invalid" class="invalid-input">
+            Please enter a valid email
+          </p>
+
+          <input type="password" class="form-input" formControlName="password" placeholder="Enter User Password" />
+          <p *ngIf="signupForm.get('password')?.touched && signupForm.get('password')?.invalid" class="invalid-input">
+            Password must be at least 6 characters
+          </p>
+
+          <button class="form-button" type="submit" [disabled]="signupForm.invalid">Sign Up</button>
+          <p>Already Have Account ? <a (click)="openLogin()">Click Here</a></p>
+        </form>
+      </div>
+    </div>
+  </div>
+
+  <!-- LOGIN -->
+  <div *ngIf="showLogin">
+    <div class="card">
+      <div class="card-header">
+        <h1>User Login</h1>
+      </div>
+      <div class="card-body">
+        <form class="common-form" [formGroup]="loginForm" (ngSubmit)="login()">
+
+          <p class="error-p">{{authError}}</p>
+
+          <input type="email" class="form-input" formControlName="email" placeholder="Enter User Email" />
+          <p *ngIf="loginForm.get('email')?.touched && loginForm.get('email')?.invalid" class="invalid-input">
+            Please enter a valid email
+          </p>
+
+          <input type="password" class="form-input" formControlName="password" placeholder="Enter User Password" />
+          <p *ngIf="loginForm.get('password')?.touched && loginForm.get('password')?.invalid" class="invalid-input">
+            Password must be at least 6 characters
+          </p>
+
+          <button class="form-button" type="submit" [disabled]="loginForm.invalid">Login</button>
+          <p>Want to Create Account ? <a (click)="openSignUp()">Click Here</a></p>
+        </form>
+      </div>
+    </div>
+  </div>
+</div>
