+++ src/app/layout/footer/footer.component.ts
