+++ src/app/layout/header/header.component.html
@@ -1,0 +1,98 @@
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark px-3">
+  <div class="container-fluid">
+    <!-- Logo -->
+    <a class="navbar-brand" routerLink="">E-Commerce</a>
+
+    <!-- Botón para responsive -->
+    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarContent" aria-controls="navbarContent" aria-expanded="false" aria-label="Toggle navigation"><span class="navbar-toggler-icon"></span>
+    </button>
+
+    <!-- Contenido principal -->
+    <div class="collapse navbar-collapse" id="navbarContent">
+      
+      <!-- Sección de Búsqueda -->
+      @if (menuType() !== 'seller') {
+        <form class="d-flex ms-auto me-3 nav-search">
+          <input  class="form-control me-2"  type="search"  #searchInput  (keyup)="searchProduct($event)"  (blur)="hideSearch()"  placeholder="Search Products" />
+
+          <button 
+            class="btn btn-outline-light" 
+            type="button" 
+            (click)="submitSearch(searchInput.value)">
+            Search
+          </button>
+        </form>
+
+        <!-- Resultados de búsqueda -->
+        @if (searchResult()) {
+          <ul class="list-group position-absolute mt-2 w-25 suggested-search">
+            @for (item of searchResult(); track item.id) {
+              <li class="list-group-item list-group-item-action">
+                <a 
+                  class="text-decoration-none" 
+                  (mousedown)="redirectToDetails(item.id)">
+                  {{ item.name }}
+                </a>
+              </li>
+            }
+          </ul>
+        }
+      }
+
+      <!-- Navegación Principal -->
+      <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
+        @switch (menuType()) {
+          @case ('default') {
+            <li class="nav-item">
+              <a class="nav-link" routerLink="seller-auth">Vendedor</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" routerLink="">Home</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" routerLink="user-auth">Ingresar</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="#">Cart({{ cartItems() }})</a>
+            </li>
+          }
+          @case ('user') { 
+            <li class="nav-item">
+              <a class="nav-link" routerLink="">Home</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link">{{ userName() | titlecase }}</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" routerLink="my-orders">Mis ordenes</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" (click)="userLogout()">Logout</a>
+            </li>
+            <li class="nav-item">
+              <a 
+                class="nav-link" 
+                [routerLink]="cartItems() > 0 ? 'cart-page' : null">
+                Cart({{ cartItems() }})
+              </a>
+            </li>
+          }
+          @case ('seller') {
+            <li class="nav-item">
+              <a class="nav-link" routerLink="create">Adicionar productos</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" routerLink="seller-home">Lista</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" (click)="logout()">Salir</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link">{{ sellerName() | titlecase }}</a>
+            </li>
+          }
+        }
+      </ul>
+    </div>
+  </div>
+</nav>
