+++ src/app/modules/my-orders/my-orders.component.ts
@@ -1,0 +1,34 @@
+import { Component, OnInit } from '@angular/core';
+import { order } from '../../data-type';
+import { ProductService } from '../../services/product.service';
+import { FormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-my-orders',standalone:true, 
+  imports:[FormsModule, CommonModule],
+  templateUrl: './my-orders.component.html',
+  styleUrls: ['./my-orders.component.css']
+})
+export class MyOrdersComponent implements OnInit {
+
+  orderData:order[]|undefined
+  constructor(private product:ProductService) { }
+
+  ngOnInit(): void {
+    this.getOrderList()
+  }
+  cancelOrder(orderId:number|undefined){
+    orderId && this.product.cancelOrder(orderId).subscribe((result)=>{
+      if(result){
+        this.getOrderList();
+      }
+    })
+  }
+  getOrderList(){
+    this.product.orderList().subscribe((result)=>{
+      this.orderData=result;
+    })
+  }
+
+}
