+++ src/app/modules/products/edit/seller-update-product.component.ts
@@ -1,0 +1,42 @@
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { product } from '../../../data-type';
+import { ProductService } from '../../../services/product.service';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+@Component({
+  selector: 'app-seller-update-product',
+  standalone:true,
+  imports:[CommonModule, FormsModule],
+  templateUrl: './seller-update-product.component.html',
+})
+export class SellerUpdateProductComponent implements OnInit {
+  productData: undefined | product;
+  productMessage: undefined | string;
+  constructor(private route: ActivatedRoute, private product: ProductService) {}
+
+  ngOnInit(): void {
+    let productId = this.route.snapshot.paramMap.get('id');
+    console.warn(productId);
+    productId &&
+      this.product.getProduct(productId).subscribe((data) => {
+        console.warn(data);
+        this.productData = data;
+      });
+  }
+  submit(data: any) {
+    if (this.productData) {
+      data.id = this.productData.id;
+    }
+    this.product.updateProduct(data).subscribe((result) => {
+      if (result) {
+        this.productMessage = 'Product has updated';
+      }
+    });
+    setTimeout(() => {
+      this.productMessage = undefined;
+    }, 3000);
+    console.warn(data);
+  }
+}
