+++ src/app/modules/seller/seller-auth/seller-auth.component.ts
@@ -1,0 +1,40 @@
+import { Component, OnInit } from '@angular/core';
+import { signUp } from '../../../data-type';
+import { SellerService } from '../../../services/seller.service';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
+
+@Component({
+  selector: 'app-seller-auth',
+  standalone: true,
+  imports: [CommonModule,FormsModule,RouterModule],
+  templateUrl: './seller-auth.component.html',
+})
+export class SellerAuthComponent implements OnInit {
+  showLogin=true;
+  authError:String='';
+  constructor(private seller: SellerService) {}
+
+  ngOnInit(): void {
+    this.seller.reloadSeller()
+  }
+  signUp(data: signUp): void {
+    console.warn(data);
+    this.seller.userSignUp(data);
+  }
+  login(data: signUp): void {
+    this.seller.userLogin(data);
+    this.seller.isLoginError.subscribe((isError)=>{
+      if(isError){
+        this.authError="Email or password is not correct";
+      }
+    })
+  }
+  openLogin(){
+    this.showLogin=true
+  }
+  openSignUp(){
+    this.showLogin=false
+  }
+}
