+++ src/app/search/search.component.html
@@ -1,11 +1,15 @@
 <div class="row search-item" *ngFor="let item of searchResult">
-    <div class="col-sm-3">
-      <a [routerLink]="['/details', item.id]">  <img src="{{item.image}}" alt=""></a>
-    </div>
-    <div class="col-sm-8 details">
-        <h6>  <a [routerLink]="['/details', item.id]"> {{item.name}} </a></h6>
-        <p><b>Price:{{item.price}}</b></p>
-        <p><span>Color:{{item.color}}</span><span>Category: {{item.category}}</span></p>
-        <p>{{item.description}}</p>
-    </div>
+  <div class="col-sm-12">
+
+<div class="card">
+  <div class="card-header"><a [routerLink]="['/details', item.id]"> <img src="{{item.image}}" alt=""></a></div>
+  <div class="card-body">
+    <h6> <a [routerLink]="['/details', item.id]"> {{item.name}} </a></h6>
+    <p><b>Price:{{item.price}}</b></p>
+    <p><span>Color:{{item.color}}</span><span>Category: {{item.category}}</span></p>
+    <p>{{item.description}}</p>
+  </div>
+</div>
+
+  </div>
 </div>
