--- src/app/seller-add-product/seller-add-product.component.css
@@ -1,7 +1,0 @@
-h1{
-    color: blueviolet;
-    text-align: center;
-}
-.add-product{
-    text-align: center;
-}
