--- src/app/seller-auth/seller-auth.component.html
@@ -1,30 +1,0 @@
-<div class="container my-5">
-  <div class="card">
-    <div class="card-header">
-      <h1 class="text-center">Seller sign-up / Registro de vendedor</h1>
-    </div>
-    <div class="card-body">
-      <div *ngIf="!showLogin" class="sign-up">
-        <form #sellerSignUp="ngForm" (ngSubmit)="signUp(sellerSignUp.value)" class="common-form">
-          <input type="text" class="form-input" placeholder="Enter Name" name="name" ngModel />
-          <input type="password" class="form-input" placeholder="Enter Password" name="password" ngModel />
-          <input type="text" class="form-input" placeholder="Enter Email" name="email" ngModel />
-          <button class="form-button">Sign Up</button>
-          <a class="auth-toggle-link" (click)="openLogin()">Already have account? Click for Login</a>
-        </form>
-      </div>
-
-      <div *ngIf="showLogin" class="login">
-        <h1 class="text-center">Seller Login</h1>
-        <p class="error-p">{{authError}}</p>
-        <form #sellerLogin="ngForm" (ngSubmit)="login(sellerLogin.value)" class="common-form">
-          <input type="text" class="form-input" placeholder="Enter Email" name="email" ngModel />
-          <input type="password" class="form-input" placeholder="Enter Password" name="password" ngModel />
-          <button class="form-button">Login</button>
-          <a class="auth-toggle-link" (click)="openSignUp()">Don't have account? Click for sign-up</a>
-
-        </form>
-      </div>
-    </div>
-  </div>
-</div>
