--- src/app/seller-home/seller-home.component.ts
@@ -1,48 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { product } from '../data-type';
-import { ProductService } from '../services/product.service';
-// import { faTrash, faEdit } from '@fortawesome/free-solid-svg-icons';
-// import { FontAwesomeModule } from '@fortawesome/angular-fontawesome'; 
-// import { IconProp } from '@fortawesome/fontawesome-svg-core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-@Component({
-  selector: 'app-seller-home',
-  standalone: true,
-  imports: [CommonModule, RouterModule],
-  templateUrl: './seller-home.component.html',
-  styleUrls: ['./seller-home.component.css'],
-})
-export class SellerHomeComponent implements OnInit {
-  
-  productList: undefined | product[];
-  productMessage: undefined | string;
-
-  constructor(private product: ProductService) {}
-
-  ngOnInit(): void {
-    this.list();
-  }
-
-  deleteProduct(id: number) {
-    this.product.deleteProduct(id).subscribe((result) => {
-      if (result) {
-        this.productMessage = 'Product is deleted';
-
-        this.list();
-      }
-    });
-    setTimeout(() => {
-      this.productMessage = undefined;
-    }, 3000);
-  }
-
-  list() {
-    this.product.productList().subscribe((result) => {
-      if (result) {
-        this.productList = result;
-      }
-    });
-  }
-}
