--- tempo/seller-add-product/seller-add-product.component.html
@@ -1,53 +1,0 @@
-<div class="add-product">
-  <h1>Add New Product</h1>
-  <p class="message-p">{{addProductMessage}}</p>
-  <form #addProduct="ngForm"
-  (ngSubmit)="submit(addProduct.value)"
-   class="common-form">
-    <input
-      type="text"
-      class="form-input"
-      placeholder="Enter Product Name"
-      name="name"
-      ngModel
-    />
-    <input
-    type="text"
-    class="form-input"
-    placeholder="Enter Product Price"
-    name="price"
-    ngModel
-  />
-  <input
-  type="text"
-  class="form-input"
-  placeholder="Enter Product Category"
-  name="category"
-  ngModel
-/>
-<input
-type="text"
-class="form-input"
-placeholder="Enter Product Color"
-name="color"
-ngModel
-/>
-<textarea
-type="text"
-class="form-input"
-placeholder="Enter Product Description"
-name="description"
-ngModel
-></textarea>
-
-<input
-type="text"
-class="form-input"
-placeholder="Enter Product Image URL"
-name="image"
-ngModel
-/>
-<button class="form-button">Add Product</button>
-
-  </form>
-</div>
