--- tempo/seller-add-product/seller-add-product.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SellerAddProductComponent } from './seller-add-product.component';
-
-describe('SellerAddProductComponent', () => {
-  let component: SellerAddProductComponent;
-  let fixture: ComponentFixture<SellerAddProductComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ SellerAddProductComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(SellerAddProductComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
