--- tempo/seller-add-product/seller-add-product.component.ts
@@ -1,28 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { product } from '../data-type';
-import { ProductService } from '../services/product.service';
-
-@Component({
-  selector: 'app-seller-add-product',
-  templateUrl: './seller-add-product.component.html',
-  styleUrls: ['./seller-add-product.component.css'],
-})
-export class SellerAddProductComponent implements OnInit {
-  addProductMessage: string | undefined;
-  constructor(private product: ProductService) {}
-
-  ngOnInit(): void {}
-
-  submit(data: product) {
-    this.product.addProduct(data).subscribe((result) => {
-      console.warn(result);
-      if (result) {
-        this.addProductMessage = 'Product is added successfully';
-      }
-    });
-
-    setTimeout(() => {
-      this.addProductMessage=undefined
-    }, 3000);
-  }
-}
